requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.11.1
    # via
    #   django
    #   uvicorn
biopython==1.79
    # via -r requirements.in
certifi==2026.4.22
    # via requests
cffi==2.0.0
    # via cryptography
chardet==3.0.4
    # via requests
click==8.3.3
    # via
    #   rq
    #   uvicorn
cryptography==39.0.1
    # via
    #   -r requirements.in
    #   http-ece
    #   py-vapid
    #   pywebpush
django==5.0.2
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-rq
django-cors-headers==3.11.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-redis-cache==3.0.0
    # via -r requirements.in
django-rq==2.5.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
http-ece==1.2.1
    # via pywebpush
httpretty==1.1.4
    # via -r requirements.in
idna==2.8
    # via requests
numpy==2.4.4
    # via biopython
psycopg2-binary==2.9.3
    # via -r requirements.in
py-vapid==1.9.2
    # via pywebpush
pycparser==3.0
    # via cffi
pywebpush==1.14.0
    # via -r requirements.in
redis==3.5.3
    # via
    #   django-redis-cache
    #   django-rq
    #   rq
requests==2.22.0
    # via
    #   -r requirements.in
    #   pywebpush
rq==1.10.1
    # via
    #   -r requirements.in
    #   django-rq
six==1.17.0
    # via pywebpush
sqlparse==0.5.5
    # via django
urllib3==1.25.11
    # via requests
uvicorn==0.17.6
    # via -r requirements.in
websockets==12.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
